+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for Ingress in `gateway-api`
 - Add support for restoring GatewayApi resources in  `gateway-api`
+- Add `noop` mode in  `gateway-api`
 
 ## v0.7 - 2022-09-27